Allowed human to play a depth limited opponent
added a lot of bug fixes and refined the valueing of the board. one
situation I ran into was when the game knows it will lose next turn (the
opponent has a guaranteed win situation) it would not create a playable
list. next steps would be making the board and length of win condition
variable size.