implemented multithreaded solution
now allows the GridPlayer to search in the background as values are
copied and displayed onto the screen. Occasionally hangs upon exit until
it is done evaluating a pattern. usually only lasts a few seconds before
it resolves. also moved definitions from GameEngine.h to be in
GameEngine.cpp, more similar refactoring to come.